# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Barrio Fino Peluqueria API",
        "description": "API documentation for Barrio Fino Peluqueria application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/citas": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get a list of all appointments",
                "operationId": "60a2d0901d3bc53094b4b617a45ac681",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "No appointments found"
                    }
                }
            }
        },
        "/api/v1/citas/store": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create a new appointment",
                "operationId": "f0534e084728287d134f5737a28132f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "barbero_id",
                                    "servicios_id",
                                    "fecha_hora",
                                    "hora_inicio",
                                    "hora_fin",
                                    "estado",
                                    "precio_total"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "barbero_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "servicios_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "fecha_hora": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-01 14:30:00"
                                    },
                                    "hora_inicio": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "14:30:00"
                                    },
                                    "hora_fin": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "15:00:00"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "pendiente"
                                    },
                                    "precio_total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/v1/citas/put/{id}": {
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update an existing appointment",
                "operationId": "aec9d6a79e8e738aec97a57641e05890",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the appointment to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "barbero_id",
                                    "servicios_id",
                                    "fecha_hora",
                                    "hora_inicio",
                                    "hora_fin",
                                    "estado",
                                    "precio_total"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "barbero_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "servicios_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "fecha_hora": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-01 14:30:00"
                                    },
                                    "hora_inicio": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "14:30:00"
                                    },
                                    "hora_fin": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "15:00:00"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "pendiente"
                                    },
                                    "precio_total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appointment updated successfully"
                    },
                    "404": {
                        "description": "Appointment not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/v1/citas/delete/{id}": {
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Delete an appointment",
                "operationId": "2ab95b2b2f6cfe470de9e0b36b197858",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the appointment to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment deleted successfully"
                    },
                    "404": {
                        "description": "Appointment not found"
                    }
                }
            }
        },
        "/api/v1/citas/searchClienteCita": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Search appointments by client name",
                "operationId": "3d872f6bf2bafe5e5e8d781d4e999c6f",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Client name to search for",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "No appointments found"
                    }
                }
            }
        },
        "/api/v1/citas/searchBarberoCita": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Search appointments by barber name",
                "operationId": "603fb9d03b81c1e974c6bf158a0483b2",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Barber name to search for",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Mike"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation"
                    },
                    "404": {
                        "description": "No appointments found"
                    }
                }
            }
        },
        "/api/v1/citas/searchFechaCita": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Search appointments by date",
                "operationId": "502c2de67ba1628e8f69dacdf959ee73",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date to search for (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-07-01"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "No appointments found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}